requirements.txt
flask
flask_cors
flask_restful
jenkins
xmltodict
mysql-connector-python